added position and move details
