+++ .github/dependabot.yml
@@ -8,7 +8,7 @@ updates:
   - package-ecosystem: "gradle" # See documentation for possible values
     directory: "/" # Location of package manifests
     schedule:
-      interval: "monthly"
+      interval: "weekly"
     commit-message:
       # Prefix all commit messages with "gradle"
       prefix: "gradle"
