+++ README.md
@@ -1,9 +1,11 @@
 # Hari: The Chess Engine
 
+Current rating: WIP
+
 ### Jump to:
 
 1. [Contributors](#contributors)
-2. [Description](#description)
+2. [Background](#background)
 3. [Using the Engine](#using-the-engine)
 4. [Architecture](#architecture)
 5. [Contributing](#contributing-to-the-project)
@@ -11,25 +13,23 @@
 
 ## Contributors
 
-1. [Divy Patel](https://github.com/divy-07) (owner)
-2. [Arnav Prasad](https://github.com/arnavpd) (owner)
+1. [Divy Patel](https://github.com/divy-07) (creator)
+2. [Arnav Prasad](https://github.com/arnavpd) (creator)
 
 Special mention: Jonathan Warkentin, some code was reused from his chess engine, Orion.
 
-## Description
+## Background
 
 [//]: # "Edit this section when implementation is complete"
 
-Note: Although the engine is currently under construction, a big chunk of the work is done.
-
-Current rating: WIP
+Note: Although the engine is still being improved, a big chunk of the work is done.
 
 The motive behind this project was to create a chess engine that can beat us, the creators, rated around 1200.
 We expect the engine to be able to beat us in the near future, once the engine is fully implemented.
 Apart from our passion for chess, we also wanted to apply the knowledge we have gained in our Software Engineering courses to a real world application.
 This includes the use of object-oriented programming, unit testing, concurrency, and other software engineering principles.
 
-The project is created using Java, as it is a nice mix between C++(speed) and Python(readability).
+The project is created using Java, as it is a nice mix between C++(speed) and Python(object-oriented).
 It is also a language that we are both familiar with.
 
 [Back to top](#hari--the-chess-engine)
@@ -49,8 +49,6 @@ Steps:
 
 ## Architecture
 
-[//]: # "Edit this section when architecture is changed"
-
 ```
              -------
              | GUI |
@@ -63,6 +61,10 @@ Steps:
             --------
             | Hari |
             --------
+                |  
+          ------------
+          | Position |
+          ------------
                 |
        ------------------
        | MoveGeneration |
@@ -74,11 +76,29 @@ Steps:
 ```
 
 The engine is composed of:
-- The UCI interface (communicates with the GUI)
-- Hari (the engine, holds the board and the game state)
-- MoveGeneration (generates the best next move)
-- Evaluation (evaluates the board position)
-- PossibleMoves (generates all possible moves for a given board position)
+- The UCI interface 
+  - communicates with the GUI
+  - handles the UCI commands from GUI
+  - sends the UCI commands from the engine to the GUI
+- Hari
+  - the engine, holds the current position
+- Position
+  - class that represents a chess position
+  - this class is immutable
+  - it contains the bitboards, the side to move, the castling rights, the en passant square, and move clocks
+- Move
+  - class that represents a chess move
+  - this class is immutable
+  - it contains the source and destination squares (as well as optional properties such as promotion piece and en passant square)
+- MoveGeneration
+  - generates the best next move
+  - currently uses the minimax algorithm
+- Evaluation
+  - evaluates the board position statically
+  - currently uses a simple material evaluation
+- PossibleMoves
+  - generates all possible moves for a given board position
+  - used to generate the moves for the minimax algorithm
 
 [Back to top](#hari--the-chess-engine)
