+++ src/main/java/chess/board/MakeMove.java
@@ -1,4 +1,4 @@
-package chess.engine;
+package chess.board;
 
 import chess.moves.Move;
